# swagger.yaml
openapi: 3.0.3
info:
  title: Branded UK E-Commerce API
  description: |
    # Overview
    This API provides endpoints for the Branded UK e-commerce platform.
    
    ## Base URL
    **Production**: `https://api.brandeduk.com`
    
    ## Authentication
    No authentication required. Public API.
    
    ## Endpoint Categories
    
    ### Products (Main)
    - `GET /api/products` - Main product listing with combined filters
    - `GET /api/products/filters` - Product type counts for filtering
    - `GET /api/products/types` - All product types with counts
    - `GET /api/products/{code}` - Single product details
    - `GET /api/products/{code}/pricing` - Get pricing information for a style code
    
    ### Contact & Quotes
    - `POST /api/contact` - Submit contact form inquiry
    - `POST /api/quotes` - Submit quote request with optional logo file uploads
    
    ### Filter Endpoints (Testing & Database Exploration)
    Each filter type has TWO endpoints:
    1. **List all values**: Returns all options from that database table
    2. **Get products by slug**: Returns products matching a specific filter value
    
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: https://api.brandeduk.com
    description: Production Server
  - url: http://localhost:3004
    description: Local Development
...
paths:
...
  /api/products:
    get:
      tags:
        - Products
      summary: Get product list (main combined filter endpoint)
      description: |
        Returns a paginated list of products with optional filtering and search.
        This is the main endpoint that supports all filters combined.
      operationId: getProducts
      parameters:
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Items per page (max 200)
          schema:
            type: integer
            default: 24
        - name: q
          in: query
          description: Search query
          schema:
            type: string
        - name: priceMin
          in: query
          schema:
            type: number
        - name: priceMax
          in: query
          schema:
            type: number
        - name: productType
          in: query
          description: Product type slug(s)
          schema:
            type: string
        - name: gender
          in: query
          schema:
            type: string
        - name: ageGroup
          in: query
          schema:
            type: string
        - name: sleeve
          in: query
          schema:
            type: string
        - name: neckline
          in: query
          schema:
            type: string
        - name: fabric
          in: query
          schema:
            type: string
        - name: size
          in: query
          schema:
            type: string
        - name: colour
          in: query
          schema:
            type: string
        - name: primaryColour
          in: query
          schema:
            type: string
        - name: style
          in: query
          schema:
            type: string
        - name: tag
          in: query
          schema:
            type: string
        - name: weight
          in: query
          schema:
            type: string
        - name: fit
          in: query
          schema:
            type: string
        - name: sector
          in: query
          schema:
            type: string
        - name: sport
          in: query
          schema:
            type: string
        - name: effect
          in: query
          schema:
            type: string
        - name: accreditations
          in: query
          schema:
            type: string
        - name: brand
          in: query
          description: Brand slug(s) or name(s) - accepts multiple values
          schema:
            type: string
          example: "awdis"
        - name: brands
          in: query
          description: Alias for brand parameter (accepts multiple values)
          schema:
            type: string
        - name: colourShade
          in: query
          schema:
            type: string
        - name: sort
          in: query
          description: |
            Sort option. Available values:
            - `newest` - Sort by creation date (newest first)
            - `best` - Best sellers (uses newest as proxy)
            - `brand-az` - Sort by brand name A-Z
            - `brand-za` - Sort by brand name Z-A
            - `code-az` - Sort by product code A-Z
            - `code-za` - Sort by product code Z-A
            - `price-lh` - Sort by price low to high
            - `price-hl` - Sort by price high to low
            - `price` - Sort by price (use with order parameter)
            - `name` - Sort by product name (use with order parameter)
          schema:
            type: string
            enum: [newest, best, brand-az, brand-za, code-az, code-za, price-lh, price-hl, price, price-low, price-high, name]
            default: newest
            default: newest
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                  page:
                    type: integer
                  limit:
                    type: integer
                  total:
                    type: integer
                  priceRange:
                    type: object
                    properties:
                      min:
                        type: number
                      max:
                        type: number
...
  /api/products/{code}:
    get:
      tags:
        - Products
      summary: Get product details by code
      operationId: getProductDetail
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
          example: "JC001"
      responses:
        '200':
          description: Successful response
        '404':
          description: Product not found
...
  /api/display-order:
    get:
      tags:
        - Display Order
      summary: List display order entries
      description: |
        Returns paginated list of `product_display_order` entries.
        You can optionally filter by `brand_id`, `product_type_id` and/or `style_code`.
      parameters:
        - name: brand_id
          in: query
          schema:
            type: integer
          description: Filter by brand ID
        - name: product_type_id
          in: query
          schema:
            type: integer
          description: Filter by product type ID
        - name: style_code
          in: query
          schema:
            type: string
          description: Filter by style code (partial match)
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/DisplayOrderEntry'
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - Display Order
      summary: Create or update a single display order entry
      description: |
        Upserts a `product_display_order` row for a specific style code and optional brand/product type context.
        Context uniqueness is enforced on `(style_code, brand_id, product_type_id)`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - style_code
                - display_order
              properties:
                style_code:
                  type: string
                  description: Product style code (e.g. "JC001")
                  example: "JC001"
                brand_id:
                  type: integer
                  nullable: true
                  description: Optional brand context ID
                  example: 28
                product_type_id:
                  type: integer
                  nullable: true
                  description: Optional product type context ID
                  example: 21
                display_order:
                  type: integer
                  description: Lower numbers appear first (1 = first)
                  example: 1
      responses:
        '201':
          description: Display order saved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Display order saved successfully
                  data:
                    $ref: '#/components/schemas/DisplayOrderEntry'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Style code not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/display-order/bulk:
    post:
      tags:
        - Display Order
      summary: Bulk update display orders in a context
      description: |
        Bulk upsert multiple display order entries in a single transaction
        for an optional `brand_id` and/or `product_type_id` context.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DisplayOrderBulkRequest'
            example:
              brand_id: 28
              product_type_id: 21
              orders:
                - style_code: "YP047"
                  display_order: 1
                - style_code: "YP049"
                  display_order: 2
      responses:
        '200':
          description: Bulk update successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Successfully updated 2 display orders
                  updated:
                    type: integer
                    example: 2
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/DisplayOrderEntry'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/display-order/{id}:
    put:
      tags:
        - Display Order
      summary: Update a display order entry by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - display_order
              properties:
                display_order:
                  type: integer
                  description: New display order value
                  example: 5
      responses:
        '200':
          description: Display order updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Display order updated successfully
                  data:
                    $ref: '#/components/schemas/DisplayOrderEntry'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Entry not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - Display Order
      summary: Delete a display order entry by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Display order deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Display order deleted successfully
                  data:
                    $ref: '#/components/schemas/DisplayOrderEntry'
        '404':
          description: Entry not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/display-order/by-context:
    delete:
      tags:
        - Display Order
      summary: Delete display order entry by style/context
      description: |
        Delete a display order row by `style_code` and optional `brand_id`/`product_type_id` context.
        If no context IDs are provided, the global (no brand/type) entry for that style will be deleted.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - style_code
              properties:
                style_code:
                  type: string
                  description: Product style code
                  example: "YP047"
                brand_id:
                  type: integer
                  nullable: true
                  example: 28
                product_type_id:
                  type: integer
                  nullable: true
                  example: 21
      responses:
        '200':
          description: Display order deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Display order deleted successfully
                  data:
                    $ref: '#/components/schemas/DisplayOrderEntry'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Entry not found for this context
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/display-order/products:
    get:
      tags:
        - Display Order
      summary: Get products with display order for admin
      description: |
        Returns all products for a brand and/or product type including their current
        `display_order` value, for use in the admin sorting interface.
      parameters:
        - name: brand_id
          in: query
          schema:
            type: integer
          description: Filter products by brand ID
        - name: product_type_id
          in: query
          schema:
            type: integer
          description: Filter products by product type ID
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/DisplayOrderProduct'
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
                  filters:
                    type: object
                    properties:
                      brand_id:
                        type: integer
                        nullable: true
                      product_type_id:
                        type: integer
                        nullable: true
        '400':
          description: Missing required brand_id or product_type_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/display-order/brands:
    get:
      tags:
        - Display Order
      summary: Get brands with product counts for display ordering
      description: Returns brands that have at least one live product, including product counts and existing brand display_order.
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/DisplayOrderBrand'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
...
    DisplayOrderEntry:
      type: object
      description: Single product_display_order row with joined metadata
      properties:
        id:
          type: integer
        style_code:
          type: string
          example: "YP047"
        brand_id:
          type: integer
          nullable: true
        brand_name:
          type: string
          nullable: true
        product_type_id:
          type: integer
          nullable: true
        product_type_name:
          type: string
          nullable: true
        display_order:
          type: integer
          description: Lower numbers appear first (1 = first)
          example: 1
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        product_name:
          type: string
          nullable: true
          description: Name of the style from styles table
...